Read Level Information From File
Changed the code so it would load the specified objects at the specified
locations, specified in each specific level file. Especially made for
the special level editor!

Doesn't load specific components with specific parameters, so for those
special cases, we still need to hardcoded those in the specific level
class. This will be included in a special future update.